refactor: Customer 입력값 검증 클래스 Domain 패키지로 재분류
